requirements.txt
kagglehub
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
pillow
opencv-python
flask
werkzeug 